refactor: shared infra, replicate and environment management
* Now the command provisioning and destroying regional replicates
patches the state and perform infrastructure changes

* The remove env CLI command now destroy cloud resources

* Shared infrastructure is now managed in a more agnostic way

The workspaces containing the "infra:deploy" and "infra:remove"
targets are considered shared-infra packages.
This way do not assume how it should be done under the hood and
which tech you should use: terraform, cloud formation, custom scripts...
The CLI will try to create a replicate of instance by environment or
use a single instance shared it accross envs acording to the config.
As a mila-runner target is used, remote caching and artifact sharing
is available

---------

Co-authored-by: Mario Arnautou <mario.arnautou@gmail.com>